// types/SalaryInput.type.ts
export type SalaryInput = {
  monthlyBasic: number;
  annualCTC: number;
  employerPFIncluded: boolean;
  standardDeduction: number;
  hra: number;
  basicDeduction80c: number;
  medicalInsurance80d: number;
  interestOnHomeLoan: number;
  nps: number;
  interestOnEducationLoan: number;
  interestOnHousingLoan: number;
  interestOnEVLoan: number;
  donation: number;
}